backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
google-cloud-firestore==2.14.0
vertexai==1.71.1

# Stable LangChain ecosystem
langchain==0.2.13
langchain-core==0.2.34
langchain-community==0.2.12
langchain-google-vertexai==1.0.10
langchain-huggingface==0.0.3

faiss-cpu==1.7.4
vaderSentiment==3.3.2
joblib==1.4.2
python-dotenv==1.0.1
